review-fix: Move HealthModal into fether-react so STATUS not passed as prop
