+++ packages/fether-ui/src/index.js
@@ -10,9 +10,7 @@ export * from './Card';
 export * from './Clickable';
 export * from './Form';
 export * from './Header';
-export * from './HealthModal';
 export * from './Modal';
 export * from './Placeholder';
-export * from './SentModal';
 export * from './TokenCard';
 export * from './utils';
